% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{ac_}
\alias{ac_}
\title{add group info for heatmap}
\usage{
ac_(expr, group_list)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{group for samples}
}
\value{
a data frame for pheatmap annotation
}
\description{
add group info for heatmap
}
