% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_hyperBar.R
\name{barplot_base}
\alias{barplot_base}
\title{barplot_base}
\usage{
barplot_base(data, bar_color, split_color)
}
\arguments{
\item{data}{a data frame from enrichGO result}

\item{bar_color}{color for bar}

\item{split_color}{color for ont}
}
\value{
ggplot ob
}
\description{
plot basic barplot for enhance barplot
}
