% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_hyperBar.R
\name{barplot_base2}
\alias{barplot_base2}
\title{barplot_base2}
\usage{
barplot_base2(data, bar_color, text_color)
}
\arguments{
\item{data}{a data frame from enrichGO result}

\item{bar_color}{color for bar}

\item{text_color}{color for text}
}
\value{
ggplot ob
}
\description{
plot basic barplot for enhance barplot not split
}
