% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{box_check}
\alias{box_check}
\title{Boxplot of expression matrix}
\usage{
box_check(
  data,
  group,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  ...
)
}
\arguments{
\item{data}{a cpm data frame of rows in genes and columns in samples}

\item{group}{a list ordered by samples in data}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}

\item{...}{more parameters in \code{\link{exprBox}}}
}
\value{
a ridges plot file
}
\description{
Check data by box plot
}
\examples{
density_check(data,group)
}
