% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanSYMBOL}
\alias{cleanSYMBOL}
\title{Convert a  data to SYMBOL named}
\usage{
cleanSYMBOL(row_counts, bad2good)
}
\arguments{
\item{row_counts}{count data frame}

\item{bad2good}{a dataframe, first column should be row name of row_counts, second should be SYMBOL}
}
\value{
a renamed dataframe
}
\description{
larger median will be retain for duplicated
}
\examples{
cleanSYMBOL(count,b2g)
}
