% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{compareEnrichCircle}
\alias{compareEnrichCircle}
\title{Plot circos of common ID}
\usage{
compareEnrichCircle(
  result_g,
  group = "model",
  height = 12000,
  width = 10000,
  mar = c(8, 0, 0, 17),
  filename = "circlize_cross.pdf",
  IDpalette = as.character(MetBrewer::met.brewer("Signac", length(unique(result_g$ID)))),
  heatCircle1 = "Count",
  heatCircle2 = "qvalue",
  heatCircle1Pal = c("#e5f5e0", "#a1d99b", "#31a354"),
  heatCircle2Pal = c("#efedf5", "#bcbddc", "#756bb1"),
  groupPal = c("#1f78b4", "#33a02c", "#ff7f00", "#b15928"),
  group_name = c("edgeR", "limma", "DESeq2", "merge")
)
}
\arguments{
\item{result_g}{results of \code{\link{modelEnrich}} or similar format data}

\item{group}{which column used to group data}

\item{height}{height of plot}

\item{width}{width of plot}

\item{mar}{margin of bottom, left, top, right around circos}

\item{filename}{filename}

\item{IDpalette}{color for unique ID}

\item{heatCircle1}{which column used to first outer circle}

\item{heatCircle2}{which column used to second outer circle}

\item{heatCircle1Pal}{color of first outer circle}

\item{heatCircle2Pal}{color of second outer circle}

\item{groupPal}{color for every group}

\item{group_name}{group name}
}
\value{
a file store plot
}
\description{
After a enrich analysis across limma, edgeR, DESeq2 results. How about the distribution among
different datset.
}
\examples{
compareEnrichCircle(result_g)
}
