% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{cor500_check}
\alias{cor500_check}
\title{Correlation of all samples and top 500 genes}
\usage{
cor500_check(
  data,
  list,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{list}{a list ordered by samples in data}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}

\item{counts_data}{a counts data frame of rows in genes and columns in samples}
}
\value{
a Heatmap shows correlation of all samples to 500 genes
}
\description{
Check data quality by calculating correlation of all samples but top 500 genes
}
\examples{
cor500_check(counts_input, list)
}
