% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_input}
\alias{counts_input}
\title{An example dataset}
\format{
An object of class \code{data.frame} with 27051 rows and 6 columns.
}
\usage{
counts_input
}
\description{
a counts data frame of rows in genes and columns in samples
}
\keyword{datasets}
