% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo-methods.R
\name{Create_dataInfo}
\alias{Create_dataInfo}
\title{dataInfo create function}
\usage{
Create_dataInfo(
  species,
  dataType,
  idType,
  expMatrix,
  groupInfo,
  caseGroup,
  filterMethod,
  matrixFiltered
)
}
\arguments{
\item{species}{species for your data. `Human` or `Mouse`}

\item{dataType}{kind of expresses value matrix. `Counts` (Integer) or `Array` (Decimal).}

\item{idType}{kind of gene id. `ENSEMBL` or `SYMBOL`, If `ENSEMBL`, it will be automatically converted to `SYMBOL`.}

\item{expMatrix}{expresses value matrix. Should be a data.frame row named by gene ID and column named by Sample}

\item{groupInfo}{a Character Vectors ordered by samples in matrix.}

\item{caseGroup}{a Character names of case group.}

\item{filterMethod}{a function used to filter expresses value matrix.}

\item{matrixFiltered}{a expresses value matrix after filter by `filterMethod`.}
}
\value{
a DEG_container ob
}
\description{
a function to create dataInfo OB
}
