% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDEGContainer.R
\name{Create_DEGContainer}
\alias{Create_DEGContainer}
\title{Create \code{DEGContainer}}
\usage{
Create_DEGContainer(
  species = "Human",
  dataType = "Counts",
  idType = "SYMBOL",
  expMatrix,
  groupInfo,
  caseGroup,
  filterMethod = "rowSums(expMatrix > 0) >= ncol(expMatrix)/2",
  msi_species = NULL,
  msi_category = "H",
  OrgDb = NULL,
  organism = NULL,
  GOTERM2GENE = NULL,
  GOTERM2NAME = NA,
  KEGGTERM2GENE = NULL,
  KEGGTERM2NAME = NA
)
}
\arguments{
\item{species}{species for your data.}

\item{dataType}{kind of expresses value matrix. `Counts` (Integer) or `Array` (Decimal).}

\item{idType}{kind of gene id. `ENSEMBL` or `SYMBOL`, If `ENSEMBL`,\code{species} supposed be one of (`Human`, `Mouse` and `Rat`), it will be automatically converted to `SYMBOL`.}

\item{expMatrix}{expresses value matrix. Should be a data.frame row named by gene ID and column named by Sample}

\item{groupInfo}{a Character Vectors ordered by samples in matrix.}

\item{caseGroup}{a Character names of case group.}

\item{filterMethod}{a function used to filter expresses value matrix. Or disable filter by set as `NULL`.}

\item{msi_species}{one of \code{\link[msigdbr]{msigdbr_species}}}

\item{msi_category}{MSigDB collection abbreviation, such as H or C1.}

\item{OrgDb}{Select an OrgDb.}

\item{organism}{Select an organism.}

\item{GOTERM2GENE}{custom enrich data of GO.}

\item{GOTERM2NAME}{custom enrich data of GO.}

\item{KEGGTERM2GENE}{custom enrich data of GSE.}

\item{KEGGTERM2NAME}{custom enrich data of GSE.}
}
\value{
a \code{DEGContainer} object
}
\description{
Begin in DEGContainer.
}
\details{
`Human`, `Mouse` and `Rat` is full supported. For other species, more setting is needed. \cr
    If data can't provided, related steps in workflow will ignore. \cr
    OrgDb can be find from here \link{https://bioconductor.org/packages/release/BiocViews.html#___OrgDb} \cr
    organism can be find from here \link{https://www.genome.jp/kegg/catalog/org_list.html} \cr
    msi_species can be get from \code{\link[msigdbr]{msigdbr_species}} \cr
}
\examples{
Create_DEGContainer(expMatrix = counts_input,groupInfo = c("C","C","C","T","T","T"),caseGroup = "T")
}
