% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_gseParam}
\alias{Create_gseParam}
\title{Create param for gse analysis}
\usage{
Create_gseParam(
  goParam = NULL,
  keggParam = NULL,
  customGO = FALSE,
  customKEGG = FALSE,
  skipGO = FALSE,
  skipKEGG = FALSE
)
}
\arguments{
\item{goParam}{arguments in \code{\link[clusterProfiler]{gseGO}} or \code{\link[clusterProfiler]{GSEA}} except \code{geneList}.}

\item{keggParam}{arguments in \code{\link[clusterProfiler]{gseKEGG}} or \code{\link[clusterProfiler]{GSEA}} except \code{geneList}.}

\item{customGO}{logic. FALSE for \code{\link[clusterProfiler]{gseGO}}, TRUE for \code{\link[clusterProfiler]{GSEA}}}

\item{customKEGG}{logic. FALSE for \code{\link[clusterProfiler]{gseKEGG}}, TRUE for \code{\link[clusterProfiler]{GSEA}}}

\item{skipGO}{logical, set parameters as NULL for skip this step}

\item{skipKEGG}{logical, set parameters as NULL for skip this step}
}
\value{
gseParam
}
\description{
Create param for gse analysis
}
\examples{
Create_gseParam()
}
