% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_msigdbGSEAparam}
\alias{Create_msigdbGSEAparam}
\title{Create \code{msigdbGSEAparam}}
\usage{
Create_msigdbGSEAparam(GSEAparam = NULL)
}
\arguments{
\item{GSEAparam}{a list contains any paramaters of \code{\link[clusterProfiler]{GSEA}} except \code{geneList}}
}
\value{
a list
}
\description{
Create a msigdbHyperParam list. If NULL, default parameters will apply.
}
\examples{
Create_msigdbGSEAparam()
}
