% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_msigdbHyperParam}
\alias{Create_msigdbHyperParam}
\title{Create \code{msigdbHyperParam}}
\usage{
Create_msigdbHyperParam(HyperParam = NULL)
}
\arguments{
\item{GSEAparam}{a list contains any paramaters of \code{\link[clusterProfiler]{enricher}} except \code{geneList}}
}
\value{
a list
}
\description{
Create a msigdbHyperParam list. If NULL, default parameters will apply.
}
\examples{
Create_msigdbHyperParam()
}
