% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_msigdbParam}
\alias{Create_msigdbParam}
\title{Create \code{MSigDB}}
\usage{
Create_msigdbParam(msigdbParam = NULL)
}
\arguments{
\item{msigdbParam}{a list contains any paramaters of \code{\link[msigdbr]{msigdbr}}}
}
\value{
a list
}
\description{
Create a msigdbParam list. If NULL, default parameters will apply.
}
