% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOB.R
\name{Create_treatInfo}
\alias{Create_treatInfo}
\title{Function for treatInfo}
\usage{
Create_treatInfo(
  cutFC = NULL,
  cutFDR = 0.05,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  sigCol = c("#2874C5", "grey", "#f87669"),
  sigAlpha = c(1, 0.5, 1),
  sigSize = c(0.6, 0.5, 0.6),
  sigShape = 16
)
}
\arguments{
\item{cutFC}{a numeric or list. Threshold of logFC for finding significantly different genes.}

\item{cutFDR}{a numeric. Threshold of Pvalue for finding significantly different genes.}

\item{label}{a character vector. Name for every group}

\item{label_ns}{a character. Name of no signification group}

\item{sigCol}{a character vector. sigColor of every group}

\item{sigAlpha}{a numeric vector. transparency for every group}
}
\value{

}
\description{
Function for treatInfo
}
