% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutFC_Verify}
\alias{cutFC_Verify}
\title{Verify the rationality of the FC threshold}
\usage{
cutFC_Verify(res, cut_FC, scale = 1)
}
\arguments{
\item{res}{one of result from limma or DESeq2 or edgeR}

\item{cut_FC}{if is null, will give a recommend value.}

\item{scale}{adjust recommend value}
}
\value{
a value
}
\description{
Verify the rationality of the FC threshold
}
\examples{
cutFC_Verify(deg_data)
}
