% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutMuch}
\alias{cutMuch}
\title{Group result of DEG analysis results}
\usage{
cutMuch(object, which, category = "H")
}
\arguments{
\item{object}{a DEGContainer}

\item{which}{kinds of DEG; can be "limma", "edgeR", "DESeq2" or "MSigDB"}

\item{category}{MSigDB collection abbreviation, such as H or C1.}
}
\value{
a DEGContainer
}
\description{
Group result of DEG analysis results
}
\examples{
cutMuch(DEGContainer,"limma","H")
}
