% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\docType{class}
\name{dataInfo-class}
\alias{dataInfo-class}
\alias{treatInfo-class}
\title{a S4 class contains inputs info for workflow}
\description{
contains begin info of RNAseqStat2

contains configurations info of RNAseqStat2
}
\seealso{
Other DEGContainer: 
\code{\link{degResults-class}}

Other DEGContainer: 
\code{\link{degResults-class}}
}
\concept{DEGContainer}
