% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEG_df}
\alias{DEG_df}
\title{An example dataset}
\format{
An object of class \code{data.frame} with 24952 rows and 6 columns.
}
\usage{
DEG_df
}
\description{
a deg data frame
}
\keyword{datasets}
