% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_here}
\alias{deg_here}
\title{Test for exists of DEG results}
\usage{
deg_here(object)
}
\arguments{
\item{object}{a DEGContainer}
}
\value{
logic vector
}
\description{
Test for exists of DEG results
}
\examples{
deg_here(DEGContainer)
}
