% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\docType{class}
\name{DEGContainer-class}
\alias{DEGContainer-class}
\title{Class \code{DEGContainer}}
\value{
The accessor functions \code{dataInfo}, \code{degResults},
  \code{hyperResults}, \code{gseResults}, \code{MSigDB} return
  the corresponding elements of a
  \code{DEGContainer}.
}
\description{
This was the universal class for storing data and
  results. We now generally reommend using the
  \code{\link{Create_DEGContainer}} to create it.
}
\section{Slots}{

\describe{
\item{\code{dataInfo}}{store input data of workflow}

\item{\code{degResults}}{store parameters and results in \code{runDEG} module}

\item{\code{hyperResults}}{store parameters and results in \code{runHyper} module}

\item{\code{gseResults}}{store parameters and results in \code{runGSEA} module}

\item{\code{MSigDB}}{store parameters and results in \code{runMSigDB} module}
}}

