% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degGroup.R
\name{degGroup}
\alias{degGroup}
\title{Group degResults}
\usage{
degGroup(obj)
}
\arguments{
\item{obj}{a DEGContainer}
}
\value{
a DEGContainer
}
\description{
group results of DEG analysis
}
\examples{
degGroup(DEGContainer)
}
