% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degResolve.R
\name{degResolve}
\alias{degResolve}
\title{run DEG minimal}
\usage{
degResolve(object, dir = ".", prefix = "2-DEG", parallel = FALSE, qc = TRUE)
}
\arguments{
\item{object}{a DEGContainer contains at least dataInfo}

\item{dir}{a directory to store results}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel execution using BiocParallel}
}
\value{
a csv file and a DEG_container ob
}
\description{
run DEG module in limma, DESeq2 and edgeR
}
\examples{
\dontrun{
degResolve(counts_input,group_list,case_group = "T", control_group = "C",dir = tempdir())
}
}
