% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R
\docType{class}
\name{degResults-class}
\alias{degResults-class}
\title{a S4 class contains degResults from deg analysis}
\description{
contains deg analysis results in RNAseqStat2
}
\seealso{
Other DEGContainer: 
\code{\link{dataInfo-class}}
}
\concept{DEGContainer}
