% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degSummary.R
\name{degSummary}
\alias{degSummary}
\title{Summary results of DEG analysis}
\usage{
degSummary(
  obj,
  dir = ".",
  prefix = "2-runDEG",
  PointVolcanoParam = list(gene = 10, light = NULL, light_color = "#24ac56",
    light_label_color = "#24ac56", expend = c(0.12, 0.12))
)
}
\arguments{
\item{obj}{a DEGContainer object}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{PointVolcanoParam}{more parameters for \code{\link{PointVolcano}}}
}
\value{
a batch of files
}
\description{
Summary results of DEG analysis
}
\examples{
degSummary(DEGContainer)
}
