% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGheatmap.R
\name{DEGtopHeatmap}
\alias{DEGtopHeatmap}
\title{Heatmap for DEG data frame}
\usage{
DEGtopHeatmap(
  object,
  which,
  top = 50,
  filename = NA,
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{object}{a counts data frame of rows in genes and columns in samples}

\item{which}{which model of deg analysis. limma edgeR or DESeq2}

\item{top}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{palette}{a color palette for plots}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}
}
\value{
a heatmap plot file
}
\description{
default will return a top 100 deg heatmap in p value = 0.05
}
\examples{
DEGtopHeatmap(object,which = "limma")
}
