% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degResolve.R
\name{edgeR_resolve}
\alias{edgeR_resolve}
\title{Basic produce of edgeR}
\usage{
edgeR_resolve(counts_data, group_list, control_group)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{control_group}{the name of the denominator level for the fold change (Control group)}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of edgeR
}
\examples{
edgeR_resolve_(counts_input, group_list, control_group= "C")
}
