% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eggnogG2T}
\alias{eggnogG2T}
\title{Convert EGGnog to GENE2TERM}
\usage{
eggnogG2T(data, SYMBOL, TERM)
}
\arguments{
\item{data}{data from eggnog}

\item{SYMBOL}{which column is SYMBOL}

\item{TERM}{which column is Pathway}
}
\value{
a data frame
}
\description{
Convert EGGnog to GENE2TERM
}
\examples{
eggnogG2T(data,SYMBOL,TERM)
}
