% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_hyperBar.R
\name{enhance_barplot}
\alias{enhance_barplot}
\title{Enhanced barplot}
\usage{
enhance_barplot(
  data,
  showCategory = 10,
  by = "Count",
  split = NULL,
  order = FALSE,
  drop = FALSE,
  fillstrip = TRUE,
  print = FALSE,
  split_color = RColorBrewer::brewer.pal(3, "Dark2"),
  bar_color = rev(RColorBrewer::brewer.pal(5, "GnBu")[3:5])
)
}
\arguments{
\item{data}{'enrichResult' object, enrichGO result}

\item{showCategory}{Category numbers to show}

\item{by}{one of Count and GeneRatio}

\item{split}{separate result by 'split' variable}

\item{order}{logical}

\item{drop}{logical}

\item{fillstrip}{fill strip rect or not}

\item{print}{logic for print plot}

\item{split_color}{color for CC, MF, and BP}

\item{bar_color}{color for bar}
}
\value{
ggplot or grid plot
}
\description{
plot enrich result in bar plot
}
\examples{
\dontrun{
test <- enrich_go(deg_data = DEG_df, x = "log2FoldChange", y = "pvalue")
enhance_barplot(test$Down,showCategory=10,split = "ONTOLOGY")
enhance_barplot(test$Down,showCategory=30)
}
}
