% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichBar.R
\name{enrichBar}
\alias{enrichBar}
\title{enrich bar plot}
\usage{
enrichBar(
  enrichResult,
  top = 10,
  group = "ONTOLOGY",
  space = 0.9,
  order_by = "pvalue",
  FDR = "qvalue",
  bar = "logP",
  bar_size = 2,
  point = "gr",
  point_shape = 21,
  point_size = 2.5,
  point_color = c("#de2d26", "#fc9272", "#fee0d2"),
  bar_color = rev(RColorBrewer::brewer.pal(5, "GnBu")[3:5]),
  group_color = RColorBrewer::brewer.pal(3, "Dark2"),
  group_names = c("Biological Process", "Cellular Component", "Molecular Function"),
  group_color_name = c("BP", "CC", "MF"),
  group_title = "ONTOLOGY",
  bar_title = "Qvalue",
  point_title = "GeneRatio",
  x_title = "-log10(pvalue)",
  y_title = "Description",
  plot_title = NULL,
  legend_text_size = 12,
  FDR_color = c("#de2d26", "#fc9272", "#fee0d2")
)
}
\arguments{
\item{enrichResult}{a \code{\link[DOSE]{enrichResult}}}

\item{top}{number of top categories to show}

\item{group}{which column used to split group}

\item{space}{space for text}

\item{order_by}{order by which column, default is pvalue}

\item{FDR}{pvalue or qvalue}

\item{bar}{which column to plot bar, default is 'logP'}

\item{bar_size}{size of bar}

\item{point}{which column to plot point, default is 'gr'}

\item{point_shape}{shape of point}

\item{point_size}{size of point}

\item{point_color}{color of point}

\item{bar_color}{color of bar}

\item{group_color}{color of group}

\item{group_names}{label of group}

\item{group_color_name}{color label of group}

\item{group_title}{title label of group}

\item{bar_title}{title of bar}

\item{point_title}{title of point}

\item{x_title}{title of x axis}

\item{y_title}{title of y axis}

\item{plot_title}{title of point}

\item{legend_text_size}{size of legend text}

\item{FDR_color}{color of FDR}
}
\value{
a ggplot
}
\description{
enrich bar plot
}
\examples{
enrichBar(enrichResult)
}
