% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-go.R
\name{enrichGO2}
\alias{enrichGO2}
\title{GO Enrichment Analysis of a gene set.}
\usage{
enrichGO2(
  gene,
  TERM2GENE,
  TERM2NAME = NA,
  organism = "UNKNOW",
  keyType = "SYMBOL",
  ont = "ALL",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  pool = FALSE
)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{TERM2GENE}{user input annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene}

\item{TERM2NAME}{user input of TERM TO NAME mapping, a data.frame of 2 column with term and name}

\item{keyType}{keytype of input gene}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant.  Tests must pass i) \code{pvalueCutoff} on unadjusted pvalues, ii) \code{pvalueCutoff} on adjusted pvalues and iii) \code{qvalueCutoff} on qvalues to be reported.}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies}
}
\value{
An \code{enrichResult} instance.
}
\description{
Given a vector of genes, this function will return the enrichment GO
categories after FDR control. \code{\link[clusterProfiler]{enrichGO}} only support OrgDb and
\code{\link[clusterProfiler]{enricher}} is too simple for GO.
In this case, a modified version of enrichGO here
}
\examples{
\dontrun{
 enrichGO2(gene,TERM2GENE,TERM2NAME)
}
}
\keyword{manip}
