% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{exprBox}
\alias{exprBox}
\title{Boxplot of all samples}
\usage{
exprBox(
  expr,
  group_list,
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  main = "Boxplot of all samples",
  filename = NA,
  x = "Samples",
  y = "log2(cpm(count)+1)",
  height = 4.46,
  width = 8.3
)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{a character vector order by samples}

\item{palette}{group palette}

\item{main}{plot title}

\item{filename}{file name you want to save plot}

\item{x}{x axis title}

\item{y}{y axis title}

\item{height}{height of plot}

\item{width}{width of plot}
}
\description{
plot a boxplot of all samples
}
\examples{
exprBox(counts_input,group_list=group_list)
}
