% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{exprHKGheatmap}
\alias{exprHKGheatmap}
\title{Human housekeeping genes Heatmap}
\usage{
exprHKGheatmap(
  expr,
  group_list,
  filename = NA,
  main = "Human housekeeping genes",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  anno_title = "Group",
  annotation_col = ac_(expr, group_list),
  width = ncol(expr) * 0.3 + 2.2,
  height = 10 * 0.3 + 2.2,
  ...,
  heatmapParam = list(show_rownames = F, cluster_cols = F)
)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{a character vector order by samples}

\item{filename}{file name you want to save plot}

\item{main}{plot title}

\item{palette}{annotation palette}

\item{anno_title}{annotation title}

\item{annotation_col}{annotation dataframe}

\item{width}{width of plot}

\item{height}{height of plot}

\item{...}{more parameters in \code{\link[pheatmap]{pheatmap}}}

\item{heatmapParam}{more parameters in \code{\link[pheatmap]{pheatmap}}}

\item{top}{if not NA, only keep top of genes by mad}
}
\value{
a heatmap or a file
}
\description{
plot a heatmap of highly uniform and strongly expressed Human housekeeping genes
}
\details{
Human housekeeping genes revisited in
https://www.tau.ac.il/~elieis/HKG/
E. Eisenberg and E.Y. Levanon, Trends in Genetics, 29 (2013)
}
\examples{
exprHKGheatmap(expr,group_list)
}
