% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{exprPCA}
\alias{exprPCA}
\title{PCA plot for Grouped Samples}
\usage{
exprPCA(
  expr,
  group_list,
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  filename = NA,
  main = "all samples - PCA",
  width = 3.5,
  height = 4,
  ...
)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{a character vector order by samples}

\item{palette}{group palette}

\item{filename}{file name you want to save plot}

\item{main}{plot title}

\item{width}{width of plot}

\item{height}{height of plot}

\item{...}{more parameters in \code{\link[factoextra]{fviz_pca_ind}}}
}
\value{
a PCA plot or a file
}
\description{
plot a pca for expression matrix
}
\details{
Use \code{\link[FactoMineR]{PCA}} to get PCA of grouped samples.
}
\examples{
exprPCA(expr,group_list)
}
