% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{exprRidges}
\alias{exprRidges}
\title{Title}
\usage{
exprRidges(
  expr,
  group_list,
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  main = "Density of gene expression",
  filename = NA,
  x = "Samples",
  y = "log2(cpm(count)+1)",
  height = 5,
  width = 8.3
)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{a character vector order by samples}

\item{palette}{group palette}

\item{main}{plot title}

\item{filename}{file name you want to save plot}

\item{x}{x axis title}

\item{y}{y axis title}

\item{height}{height of plot}

\item{width}{width of plot}
}
\value{
a Ridges plot or a file
}
\description{
Title
}
\examples{
exprRidges(expr,group_list)
}
