% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_check.R
\name{exprTopHeatmap}
\alias{exprTopHeatmap}
\title{Expression Heatmap}
\usage{
exprTopHeatmap(
  expr,
  group_list,
  filename = NA,
  top = 1000,
  main = "SD Top 1000 genes",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  anno_title = "Group",
  annotation_col = ac_(expr, group_list),
  width = ncol(expr) * 0.3 + 2.2,
  height = ncol(expr) * 0.3 + 2.2,
  ...,
  heatmapParam = list(show_rownames = F)
)
}
\arguments{
\item{expr}{a expression matrix}

\item{group_list}{a character vector order by samples}

\item{filename}{file name you want to save plot}

\item{top}{if not NA, only keep top of genes by mad}

\item{main}{plot title}

\item{palette}{annotation palette}

\item{anno_title}{annotation title}

\item{annotation_col}{annotation dataframe}

\item{width}{width of plot}

\item{height}{height of plot}

\item{...}{more parameters in \code{\link[pheatmap]{pheatmap}}}

\item{heatmapParam}{more parameters in \code{\link[pheatmap]{pheatmap}}}
}
\value{
a heatmap or a file
}
\description{
plot a heatmap of expression matrix
}
\details{
Can filter genes by `top`.
`top` means the top sd (by \code{\link[stats]{sd}}) of expression matrix.
}
\examples{
exprTopHeatmap(expr,group_list)
}
