% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FC_Identify}
\alias{FC_Identify}
\title{Identify the column where the FC is located}
\usage{
FC_Identify(res)
}
\arguments{
\item{res}{one of result from limma or DESeq2 or edgeR}
}
\value{
a column name
}
\description{
Identify the column where the FC is located
}
\examples{
FC_Identify(deg_data)
}
