% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGvolcano.R
\name{geom_volcano_point}
\alias{geom_volcano_point}
\title{A preset geom of \code{\link[ggplot2]{geom_point}}}
\usage{
geom_volcano_point(
  data,
  shape = 1,
  stroke = 0.15,
  color = "#24ac56",
  fill = NA,
  ...,
  mapping = NULL
)
}
\value{
geom
}
\description{
A preset geom of \code{\link[ggplot2]{geom_point}}
}
