% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGvolcano.R
\name{geom_volcano_text}
\alias{geom_volcano_text}
\title{A preset geom of \code{\link[ggrepel]{geom_text_repel}}}
\usage{
geom_volcano_text(
  data,
  mapping = NULL,
  nudge_x = NULL,
  nudge_y = -1,
  hjust = 0,
  size = 1.8,
  direction = "y",
  segment.size = 0.1,
  segment.linetype = 6,
  max.overlaps = 10,
  max.iter = 1e+06,
  max.time = 10,
  min.segment.length = 0,
  fontface = "bold",
  family = "Times",
  ...
)
}
\value{
geom
}
\description{
A preset geom of \code{\link[ggrepel]{geom_text_repel}}
}
