% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseSummary.R
\name{gesSummary}
\alias{gesSummary}
\title{Summary GSE results}
\usage{
gesSummary(obj, dir = ".", prefix = "4-runGSEA", top = 10)
}
\arguments{
\item{obj}{a DEGContainer}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top rows of up and down}
}
\value{
files
}
\description{
Summary GSE results
}
\examples{
gesSummary(DEGContainer)
}
