% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGvolcano.R
\name{get_breaks}
\alias{get_breaks}
\title{Values break parameters \code{\link[ggplot2]{scale_colour_identity}}}
\usage{
get_breaks(volcano, column_name)
}
\arguments{
\item{volcano}{data from \code{\link{create_volcano}}}

\item{column_name}{a column name, e.g. "point.color"}
}
\value{
Values for break parameters
}
\description{
Values break parameters \code{\link[ggplot2]{scale_colour_identity}}
}
