% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_res}
\alias{go_res}
\title{An example dataset}
\format{
An object of class \code{data.frame} with 270 rows and 10 columns.
}
\usage{
go_res
}
\description{
a go result data frame
}
\keyword{datasets}
