% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_list}
\alias{group_list}
\title{An example dataset}
\format{
An object of class \code{character} of length 6.
}
\usage{
group_list
}
\description{
a list ordered by samples in counts_data
}
\keyword{datasets}
