% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{groupCir}
\alias{groupCir}
\title{Group circlize matrix}
\usage{
groupCir(matrix)
}
\arguments{
\item{matrix}{row names is from, column names is to}
}
\value{
a named character vector
}
\description{
Group circlize matrix
}
\examples{
groupCir(matrix)
}
