% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_gse.R
\name{GSEAbar}
\alias{GSEAbar}
\title{gseKEGG bar plot}
\usage{
GSEAbar(res, top = 10)
}
\arguments{
\item{top}{top rows of up and down}

\item{data}{a enrich_gsekegg result}

\item{pvalue_cut}{filter by pvalue}

\item{enrichmentScore_cut}{filter by enrichmentScore}
}
\value{
ggplot ob
}
\description{
plot for enrich_gsekegg result
}
\examples{
\dontrun{
test <- enrich_gsekegg(DEG_df,x = "log2FoldChange")
geskegg_barplot(test)
}
}
