% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_gse.R
\name{GSEAplot}
\alias{GSEAplot}
\title{plot gseKEGG plot}
\usage{
GSEAplot(res, top = 10)
}
\arguments{
\item{res}{output from gse}

\item{top}{filter top by pvalue}

\item{pvalue_cut}{filter cut of pvalue}

\item{enrichmentScore_cut}{filter cut of enrichmentScore}
}
\value{
a list contains up and down
}
\description{
plot gseKEGG plots by  up and down
}
\examples{
\dontrun{
gsekegg_res <- enrich_gsekegg(DEG_df,x = "log2FoldChange")
plots_l <- enhance_gseplot(gsekegg_res)
}
}
