% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbResolve.R
\name{gseMSigDB}
\alias{gseMSigDB}
\title{GSE analysis of MSigDB data sets}
\usage{
gseMSigDB(object)
}
\arguments{
\item{object}{a DEGContainer}
}
\value{
a DEGContainer
}
\description{
GSE analysis of MSigDB data sets
}
\examples{
gseMSigDB(DEGContainer)
}
