% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseResolve.R
\name{gseResolve}
\alias{gseResolve}
\title{GSEA analysis}
\usage{
gseResolve(object, GO = FALSE, KEGG = TRUE)
}
\arguments{
\item{object}{a DEGContainer}

\item{GO}{run GO database}

\item{KEGG}{run KEGG database}
}
\value{
a DEGContainer
}
\description{
GSEA analysis
}
\examples{
gseResolve(DEGContainer)
}
