% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbResolve.R
\name{gsvaResolve}
\alias{gsvaResolve}
\title{GSVA analysis}
\usage{
gsvaResolve(object)
}
\arguments{
\item{object}{a DEGContainer}
}
\value{

}
\description{
GSVA analysis
}
\examples{
gsvaResolve(DEGContainer)
}
