% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{heatLegend}
\alias{heatLegend}
\title{Legend for heatmap}
\usage{
heatLegend(
  annoData,
  palatte,
  ...,
  title = "Values",
  title_position = "topleft",
  int = T
)
}
\arguments{
\item{annoData}{a}

\item{palatte}{p}

\item{...}{more Legend parameters}

\item{title}{t}

\item{title_position}{t}

\item{int}{integer value?}
}
\value{
a legend of circlize
}
\description{
Legend for heatmap
}
\examples{
heatLegend(annoData, palatte)
}
