% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{highlightGroup}
\alias{highlightGroup}
\title{Group highlight}
\usage{
highlightGroup(
  df,
  track.index = 1,
  palatte,
  cex = 1,
  text.col = "white",
  niceFacing = TRUE,
  text.vjust = "4mm",
  ...
)
}
\arguments{
\item{df}{d}

\item{track.index}{t}

\item{palatte}{p}

\item{cex}{c}

\item{text.col}{t}

\item{niceFacing}{f}

\item{text.vjust}{t}

\item{...}{more highlight.sector parameters}
}
\value{
a highlight of circlize
}
\description{
Group highlight
}
\examples{
highlightGroup(df,track.index = 1,palatte)
}
