% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{HKG_check}
\alias{HKG_check}
\title{Human housekeeping genes expression}
\usage{
HKG_check(
  data,
  group,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{data}{a cpm data frame of rows in genes and columns in samples}

\item{group}{a list ordered by samples in data}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}
}
\value{
a heatmap plot file
}
\description{
Check Human housekeeping genes expression
}
\examples{
HKG_check(data,group)
}
