% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hyper_GS}
\alias{hyper_GS}
\title{Get a significant difference gene list}
\usage{
hyper_GS(object, which, type, OrgDb = NULL, category = "H")
}
\arguments{
\item{object}{a DEGContainer}

\item{which}{kinds of DEG; can be "limma", "edgeR", "DESeq2" or "MSigDB"}

\item{type}{"ENTREZID" or "SYMBOL"}

\item{OrgDb}{OrgDb}

\item{category}{MSigDB collection abbreviation, such as H or C1.}
}
\value{

}
\description{
Get a significant difference gene list
}
\examples{
hyper_GS(object, which, type)
}
