% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichBar.R
\name{hyperBar}
\alias{hyperBar}
\title{hyper barplot}
\usage{
hyperBar(res, top = 10)
}
\arguments{
\item{res}{a \code{\link[DOSE]{enrichResult}}}

\item{top}{number of top categories to show, by pvalue}
}
\value{

}
\description{
hyper barplot
}
\examples{
hyperBar(res)
}
