% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbResolve.R
\name{hyperMSigDB}
\alias{hyperMSigDB}
\title{Hyper analysis of DEGContainer}
\usage{
hyperMSigDB(object)
}
\arguments{
\item{object}{a DEGContainer}
}
\value{
a DEGContainer
}
\description{
Hyper analysis of DEGContainer
}
\examples{
hyperMSigDB(DEGContainer)
}
