% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbResolve.R
\name{hyperMSigDB_Resolve}
\alias{hyperMSigDB_Resolve}
\title{Hyper analysis for MSigDB}
\usage{
hyperMSigDB_Resolve(..., geneSet_list, msigdbHyperParam)
}
\arguments{
\item{...}{more parameters for \code{\link[clusterProfiler]{enricher}}}

\item{geneSet_list}{a gene set list}

\item{msigdbHyperParam}{parameters for hyper}
}
\value{
a list of ernrichResult
}
\description{
Hyper analysis for MSigDB
}
\examples{
hyperMSigDB_Resolve()
}
