% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperResolve.R
\name{hyperResolve}
\alias{hyperResolve}
\title{Resolve Hyper}
\usage{
hyperResolve(object, GO = FALSE, KEGG = TRUE)
}
\arguments{
\item{object}{DEGContainer}

\item{GO}{run GO,TRUE or FALSE}

\item{KEGG}{run KEGG,TRUE or FALSE}
}
\value{
DEGContainer
}
\description{
Resolve Hyper
}
\examples{
hyperResolve(object, GO = FALSE, KEGG = TRUE)
}
