% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperSummary.R
\name{hyperSummary}
\alias{hyperSummary}
\title{Summary Hyper results}
\usage{
hyperSummary(obj, dir = ".", prefix = "3-runHyper", top = 10)
}
\arguments{
\item{obj}{a DEGContainer}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top rows of up and down}
}
\value{
files
}
\description{
Summary Hyper results
}
\examples{
hyperSummary(DEGContainer)
}
