% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{idLegend}
\alias{idLegend}
\title{Legend for ID}
\usage{
idLegend(
  df,
  palatte,
  type = "points",
  title = "Pathways",
  title_position = "topleft",
  ...
)
}
\arguments{
\item{df}{df}

\item{palatte}{p}

\item{type}{t}

\item{title}{t}

\item{title_position}{t}

\item{text_width}{t}
}
\value{
a legend of circlize
}
\description{
Legend for ID
}
\examples{
idLegend(df, palatte)
}
