% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degResolve.R
\name{limma_resolve}
\alias{limma_resolve}
\title{Basic produce of limma}
\usage{
limma_resolve(counts_data, group_list, control_group, case_group)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{control_group}{the name of the denominator level for the fold change (Control group)}

\item{case_group}{the name of the numerator level for the fold change (Test group)}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of limma
}
\examples{
limma_resolve(counts_input, group_list, control_group= "C", case_group = "T")
}
