% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{linkCir}
\alias{linkCir}
\title{Create matrix for circlize chordDiagram}
\usage{
linkCir(df, group = "model")
}
\arguments{
\item{df}{a data frame contains multigroup results of enrich}

\item{group}{which column is group data}
}
\value{
a matrix contains relations of IDs in group
}
\description{
Create matrix for circlize chordDiagram
}
\examples{
linkCir(df)
}
