% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{modelEnrich}
\alias{modelEnrich}
\title{Get circlize data from obj after hyper analysis}
\usage{
modelEnrich(obj, dataBase, orderBy = "pvalue", head = 3)
}
\arguments{
\item{obj}{obj}

\item{dataBase}{KEGG or GO}

\item{orderBy}{pvalue or other column}

\item{head}{head after order}
}
\value{
a list
}
\description{
Get circlize data from obj after hyper analysis
}
\examples{
modelEnrich(obj)
}
