% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R, R/allMethods.R
\docType{methods}
\name{MSigDB}
\alias{MSigDB}
\alias{MSigDB-class}
\alias{msigdbParam,MSigDB-method}
\alias{msigdbGSEAparam,MSigDB-method}
\alias{msigdbHyperParam,MSigDB-method}
\alias{msigdbParam<-,MSigDB-method}
\alias{msigdbGSEAparam<-,MSigDB-method}
\alias{msigdbHyperParam<-,MSigDB-method}
\alias{msigdbData,MSigDB-method}
\alias{msigdbData<-,MSigDB-method}
\alias{msigdbGSEAresult,MSigDB-method}
\alias{msigdbGSEAresult<-,MSigDB-method}
\alias{msigdbHyperResult,MSigDB-method}
\alias{msigdbHyperResult<-,MSigDB-method}
\alias{msigdbGSVAresult,MSigDB-method}
\alias{msigdbGSVAresult<-,MSigDB-method}
\alias{msigdbTreat,MSigDB-method}
\alias{msigdbTreat<-,MSigDB-method}
\title{Class \code{MSigDB}}
\usage{
\S4method{msigdbParam}{MSigDB}(obj)

\S4method{msigdbGSEAparam}{MSigDB}(obj)

\S4method{msigdbHyperParam}{MSigDB}(obj)

\S4method{msigdbParam}{MSigDB}(obj) <- value

\S4method{msigdbGSEAparam}{MSigDB}(obj) <- value

\S4method{msigdbHyperParam}{MSigDB}(obj) <- value

\S4method{msigdbData}{MSigDB}(obj)

\S4method{msigdbData}{MSigDB}(obj) <- value

\S4method{msigdbGSEAresult}{MSigDB}(obj)

\S4method{msigdbGSEAresult}{MSigDB}(obj) <- value

\S4method{msigdbHyperResult}{MSigDB}(obj)

\S4method{msigdbHyperResult}{MSigDB}(obj) <- value

\S4method{msigdbGSVAresult}{MSigDB}(obj)

\S4method{msigdbGSVAresult}{MSigDB}(obj) <- value

\S4method{msigdbTreat}{MSigDB}(obj)

\S4method{msigdbTreat}{MSigDB}(obj) <- value
}
\value{
The accessor functions \code{msigdbParam}, \code{msigdbData},
  \code{msigdbGSEAparam}, \code{msigdbGSEAresult}, \code{msigdbGSVAresult}
  return the corresponding elements of a
  \code{DEGContainer} or \code{MSigDB}.
}
\description{
This was the class for storing data from MSigDB and analysis
  results.
  We now generally recommend using the \code{\link{Create_DEGContainer}}
  to create it in \code{DEGContainer} obj.
}
\section{Slots}{

\describe{
\item{\code{msigdbParam}}{list_OR_NULL. Store Param of \code{\link[msigdbr]{msigdbr}}}

\item{\code{msigdbData}}{list. Save data download from MSigDB by \code{\link[msigdbr]{msigdbr}}.}

\item{\code{msigdbGSEAparam}}{list_OR_NULL. Store Param of \code{\link[clusterProfiler]{GSEA}}}

\item{\code{msigdbHyperParam}}{list_OR_NULL. Store Param of \code{\link[clusterProfiler]{enricher}}}

\item{\code{msigdbGSEAresult}}{list. Store results of GSEA.}

\item{\code{msigdbHyperResult}}{list. Store results of Hyper}

\item{\code{msigdbGSVAresult}}{list. Store results of GSVA.}
}}

\seealso{
Other DEGContainer: 
\code{\link{dataInfo}},
\code{\link{degResults-class}}
}
\concept{DEGContainer}
