% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbResolve.R
\name{msigdbGet}
\alias{msigdbGet}
\title{Download MSigDB data for DEGContainer}
\usage{
msigdbGet(object)
}
\arguments{
\item{object}{a DEGContainer}
}
\value{
a DEGContainer contains MSigDB data
}
\description{
Download MSigDB data for DEGContainer
}
\examples{
msigdbGet(degcontainer)
}
