% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigDBSummary.R
\name{MSigDBSummary}
\alias{MSigDBSummary}
\title{Summary MSigDB modules results}
\usage{
MSigDBSummary(obj, dir = ".", prefix = "5-runMSigDB", top = 10)
}
\arguments{
\item{obj}{a DEGContainer}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top of hyper or gsea}
}
\value{

}
\description{
Summary MSigDB modules results
}
\examples{
MSigDBSummary(data_msigdb)
}
