% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOB.R
\docType{methods}
\name{newDEGContainer}
\alias{newDEGContainer}
\title{Initialize an object of class \code{DEGContainer}}
\usage{
newDEGContainer(
  dataInfo = NULL,
  degResults = NULL,
  hyperResults = NULL,
  gseResults = NULL,
  MSigDB = NULL
)
}
\arguments{
\item{dataInfo}{a dataInfo object}

\item{degResults}{a degResults object}

\item{hyperResults}{a hyperResults object}

\item{gseResults}{a gseResults object}

\item{MSigDB}{a MSigDB object}
}
\value{
a DEGContainer
}
\description{
Constructs a \code{DEGContainer} object. Additional helper
  methods for manipulating \code{DEGContainer} objects are  also
  described below. We now recommend using
  \code{Create_DEGContainer} to create objects, instead.
}
