% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{paletteCir}
\alias{paletteCir}
\title{Create a palette of ID across different group}
\usage{
paletteCir(
  df,
  group = "model",
  palette = as.character(MetBrewer::met.brewer("Signac", length(unique(df$ID))))
)
}
\arguments{
\item{df}{a data frame contains multigroup results of enrich}

\item{group}{which column is group data}

\item{palette}{a color vecter, should be length with unique ID}
}
\value{
a named palette
}
\description{
Create a palette of ID across different group
}
\examples{
paletteCir(df)
}
