% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{pca_check}
\alias{pca_check}
\title{PCA QC check}
\usage{
pca_check(
  data,
  list,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{data}{a cpm data frame of rows in genes and columns in samples}

\item{list}{a list ordered by samples in data}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}
}
\value{
a figure of PCA
}
\description{
check data quality by PCA plot
}
\examples{
pca_check(data, list)
}
