% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGvolcano.R
\name{PointVolcano}
\alias{PointVolcano}
\title{Showing gene labels and highlighting gene points in a volcano plot in ggplot}
\usage{
PointVolcano(
  object,
  which,
  category = "H",
  gene,
  light = NULL,
  light_color = "#24ac56",
  light_label_color = "#24ac56",
  expend = c(0.12, 0.12),
  nudge_y_up = -6.5,
  nudge_y_down = -4
)
}
\arguments{
\item{object}{an DEGContainer}

\item{which}{limma edgeR or DESeq2}

\item{category}{MSigDB collection abbreviation, such as H or C1.}

\item{gene}{number or gene name}

\item{light}{number or gene name}

\item{light_color}{character}

\item{light_label_color}{character}

\item{expend}{c(0.12, 0.12)}

\item{nudge_y_up}{nudge up gene label in y direction}

\item{nudge_y_down}{nudge down gene label in y direction}
}
\value{
a ggplot ob
}
\description{
specific some genes to label
}
\examples{
PointVolcano(DEGContainer,"limma")
}
