% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvalue_Identify}
\alias{pvalue_Identify}
\title{Identify the column where the pvalue is located}
\usage{
pvalue_Identify(res)
}
\arguments{
\item{res}{one of result from limma or DESeq2 or edgeR}
}
\value{
a column name
}
\description{
Identify the column where the pvalue is located
}
\examples{
pvalue_Identify(deg_data)
}
