% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{rectCir}
\alias{rectCir}
\title{Plot rect in continuous value}
\usage{
rectCir(
  preData,
  annoData,
  palatte,
  track.index = 1,
  ylim = c(0, 1),
  y1Scale = 0,
  y2Scale = 0.5
)
}
\arguments{
\item{preData}{data after pre plot}

\item{annoData}{annotation data}

\item{palatte}{palatte}

\item{track.index}{index number}

\item{ylim}{y range}

\item{y1Scale}{adjust y start}

\item{y2Scale}{adjust y end}
}
\value{
a rect plot of circlize
}
\description{
Plot rect in continuous value
}
\examples{
rectCir(preData, annoData, palatte)
}
