% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichCirclize.R
\name{rectCirDiscrete}
\alias{rectCirDiscrete}
\title{plot rect in discrete value}
\usage{
rectCirDiscrete(
  preData,
  annoData,
  palatte,
  track.index = 1,
  ylim = c(0, 1),
  y1Scale = 0,
  y2Scale = 0.5
)
}
\arguments{
\item{preData}{preData}

\item{annoData}{annoData}

\item{palatte}{palatte}

\item{track.index}{track.index}

\item{ylim}{ylim}

\item{y1Scale}{y1Scale}

\item{y2Scale}{y2Scale}
}
\value{
a rect plot of circlize
}
\description{
plot rect in discrete value
}
\examples{
rectCirDiscrete(preData, annoData, palatte)
}
