% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runALL.R
\name{runALL}
\alias{runALL}
\title{run all workflow}
\usage{
runALL(object, dir = "output", top = 10, parallel = T, GO = FALSE, KEGG = TRUE)
}
\arguments{
\item{object}{a DEGContainer}

\item{dir}{a directory to store results}

\item{top}{top for enrich result filter}

\item{parallel}{use parallel in DESeq2}

\item{GO}{logic, run GO in enrich step}

\item{KEGG}{logic, run KEGG in enrich step}
}
\value{
a DEGContainer
}
\description{
runALL include runCheck runDEG runHyper runGSEA runMSigDB
}
