% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{runCheck}
\alias{runCheck}
\title{Check you data by PCA and Heatmaps}
\usage{
runCheck(
  object,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{object}{a DEGContainer contains at least dataInfo}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}
}
\description{
A previous check function for overview the data sets.
}
\examples{
runCheck(object, tempdir())
}
