% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSEA.R
\name{runGSEA}
\alias{runGSEA}
\title{runGSEA Module}
\usage{
runGSEA(
  obj,
  dir = ".",
  prefix = "4-runGSEA",
  top = 10,
  GO = FALSE,
  KEGG = TRUE
)
}
\arguments{
\item{obj}{a DEGContainer}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top rows of up and down}

\item{GO}{run GO}

\item{KEGG}{run KEGG}
}
\value{

}
\description{
runGSEA Module control GO and KEGG
}
\examples{
runGSEA(DEGContainer)
}
