% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHyper.R
\name{runHyper}
\alias{runHyper}
\title{runHyper}
\usage{
runHyper(
  obj,
  dir = ".",
  prefix = "3-runHyper",
  top = 10,
  GO = FALSE,
  KEGG = TRUE
)
}
\arguments{
\item{obj}{a DEGContainer}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top for enrich result filter}

\item{GO}{logic, run GO in enrich step}

\item{KEGG}{logic, run KEGG in enrich step}
}
\value{
a DEGContainer
}
\description{
runHyper
}
\examples{
runHyper(DEGContainer)
}
