% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSigDB.R
\name{runMSigDB}
\alias{runMSigDB}
\title{run MSigDB module}
\usage{
runMSigDB(obj, dir = ".", prefix = "5-runMSigDB", top = 10)
}
\arguments{
\item{obj}{a DEGContainer after MSigDB analysis}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{top}{top}
}
\value{

}
\description{
run MSigDB module
}
\examples{
runMSigDB(data_gse)
}
