% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_enrichBar.R
\name{textBarData}
\alias{textBarData}
\title{shape data of enrichResult}
\usage{
textBarData(
  enrichResult,
  richFactor = FALSE,
  top = 10,
  group = "ONTOLOGY",
  order_by = "pvalue"
)
}
\arguments{
\item{enrichResult}{a \code{\link[DOSE]{enrichResult}}}

\item{richFactor}{need to add richFactor?}

\item{top}{number of top categories to show, by order_by}

\item{group}{group data in which column}

\item{order_by}{order data by which column}
}
\description{
shape data of enrichResult
}
\examples{
textBarData(enrichResult)
}
