% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEGvolcano.R
\name{theme_volcano}
\alias{theme_volcano}
\title{theme_volcano}
\usage{
theme_volcano(...)
}
\arguments{
\item{...}{param passed from theme}
}
\value{
a ggplot theme
}
\description{
a nice theme for DEG volcano
}
