% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCheck.R
\name{top1000_check}
\alias{top1000_check}
\title{Heatmap of all samples and Top1000 genes}
\usage{
top1000_check(
  data,
  group,
  dir = ".",
  prefix = "1-run_check",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{data}{a cpm data frame of rows in genes and columns in samples}

\item{group}{a list ordered by samples in data}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}
}
\value{
a Heatmap shows top100 genes of all samples
}
\description{
Check data quality by plotting heatmap of top 1000 genes
}
\examples{
top1000_check(data, group)
}
