% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{topGene}
\alias{topGene}
\title{Choose top genes from DEG data frame of limma DESeq2 edgeR}
\usage{
topGene(object, topSig = 50, which, category = "H")
}
\arguments{
\item{object}{a DEG data frame contains logFC and p value}

\item{topSig}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{which}{a single number for threshold value of P value}
}
\value{
a character vector of top genes
}
\description{
For example, if topSig = 50, head 50 and tail 50 will return from a
ordered DEG data frame after filtering by threshold value of P value
}
\examples{
topGene(object,  topSig = 50, which = "limma")
}
