% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toSYMBOL}
\alias{toSYMBOL}
\title{Convert a ENSEMBL named data to SYMBOL named}
\usage{
toSYMBOL(row_counts, species)
}
\arguments{
\item{row_counts}{count data frame}

\item{species}{support Human Mouse Rat}
}
\value{
a renamed dataframe
}
\description{
Convert a ENSEMBL named data to SYMBOL named
}
\examples{
toSYMBOL()
}
