% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGContainer-methods.R
\name{Create_DEGContainer}
\alias{Create_DEGContainer}
\title{Create \code{DEGContainer}}
\usage{
Create_DEGContainer(
  species = "Human",
  dataType = "Counts",
  idType = "SYMBOL",
  expMatrix,
  groupInfo,
  caseGroup,
  filterMethod = "rowSums(expMatrix > 0) >= ncol(expMatrix)/2"
)
}
\arguments{
\item{species}{species for your data. `Human` or `Mouse`.}

\item{dataType}{kind of expresses value matrix. `Counts` (Integer) or `Array` (Decimal).}

\item{idType}{kind of gene id. `ENSEMBL` or `SYMBOL`, If `ENSEMBL`, it will be automatically converted to `SYMBOL`.}

\item{groupInfo}{a Character Vectors ordered by samples in matrix.}

\item{caseGroup}{a Character names of case group.}

\item{filterMethod}{a function used to filter expresses value matrix. Or disable filter by set as `NULL`.}

\item{matrix}{expresses value matrix. Should be a data.frame row named by gene ID and column named by Sample}
}
\value{
a \code{DEGContainer} object
}
\description{
Begin in DEGContainer.
}
\examples{
Create_DEGContainer(expMatrix = counts_input,groupInfo = c("C","C","C","T","T","T"),caseGroup = "T")
}
