% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_hyperParam}
\alias{Create_hyperParam}
\title{Create param for hyper analysis}
\usage{
Create_hyperParam(
  goParam = NULL,
  keggParam = NULL,
  customGO = FALSE,
  customKEGG = FALSE,
  skipGO = FALSE,
  skipKEGG = FALSE
)
}
\arguments{
\item{goParam}{arguments in \code{\link[clusterProfiler]{enrichGO}} or \code{\link[clusterProfiler]{enricher}} except \code{gene}.}

\item{keggParam}{arguments in \code{\link[clusterProfiler]{enrichKEGG}} or \code{\link[clusterProfiler]{enricher}} except \code{gene}.}

\item{customGO}{logic. FALSE for \code{\link[clusterProfiler]{enrichGO}}, TRUE for \code{\link[clusterProfiler]{enricher}}}

\item{customKEGG}{logic. FALSE for \code{\link[clusterProfiler]{enrichKEGG}}, TRUE for \code{\link[clusterProfiler]{enricher}}}

\item{skipGO}{logical, set parameters as NULL for skip this step}

\item{skipKEGG}{logical, set parameters as NULL for skip this step}
}
\value{
hyperParam
}
\description{
Create param for hyper analysis
}
\examples{
Create_hyperParam()
}
