% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{Create_MSigDB}
\alias{Create_MSigDB}
\title{Create \code{MSigDB}}
\usage{
Create_MSigDB(
  msigdbParam = Create_msigdbParam(),
  msigdbGSEAparam = Create_msigdbGSEAparam()
)
}
\arguments{
\item{msigdbParam}{from \code{Create_msigdbParam}}

\item{msigdbGSEAparam}{from \code{Create_msigdbGSEAparam}}
}
\value{
a \code{MSigDB} object
}
\description{
Create a MSigDB object.
}
