% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParam.R
\name{Create_msigdbGSEAparam}
\alias{Create_msigdbGSEAparam}
\title{Create \code{MSigDB}}
\usage{
Create_msigdbGSEAparam(GSEAparam = NULL)
}
\arguments{
\item{GSEAparam}{a list contains any paramaters of \code{\link[clusterProfiler]{GSEA}}}
}
\value{
a list
}
\description{
Create a GSEAparam list. If NULL, default parameters will apply.
}
