% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R, R/allGeneric.R
\docType{methods}
\name{dataInfo}
\alias{dataInfo}
\alias{dataInfo-class}
\alias{treatInfo-class}
\alias{species}
\alias{dataType}
\alias{idType}
\alias{expMatrix}
\alias{groupInfo}
\alias{caseGroup}
\alias{filterMethod}
\alias{matrixFiltered}
\alias{geneNames}
\alias{sampleNames}
\alias{species<-}
\alias{dataType<-}
\alias{idType<-}
\alias{expMatrix<-}
\alias{groupInfo<-}
\alias{caseGroup<-}
\alias{filterMethod<-}
\alias{matrixFiltered<-}
\alias{geneNames<-}
\alias{sampleNames<-}
\title{Class \code{dataInfo}}
\usage{
species(obj)

dataType(obj)

idType(obj)

expMatrix(obj)

groupInfo(obj)

caseGroup(obj)

filterMethod(obj)

matrixFiltered(obj)

geneNames(obj, filtered)

sampleNames(obj, filtered)

species(obj) <- value

dataType(obj) <- value

idType(obj) <- value

expMatrix(obj) <- value

groupInfo(obj) <- value

caseGroup(obj) <- value

filterMethod(obj) <- value

matrixFiltered(obj) <- value

geneNames(obj, filtered) <- value

sampleNames(obj, filtered) <- value
}
\value{
The accessor functions \code{species}, \code{dataType},
  \code{idType}, \code{expMatrix}, \code{groupInfo},
  \code{caseGroup}, \code{filterMethod}, \code{matrixFiltered}
  return the corresponding elements of a
  \code{DEGContainer} or \code{dataInfo}.
}
\description{
This was the class for storing input data for workflow.
  We now generally recommend using the \code{\link{Create_DEGContainer}}
  to create it in \code{DEGContainer} obj.

contains configurations info of RNAseqStat2
}
\section{Slots}{

\describe{
\item{\code{species}}{species for your data. `Human` or `Mouse`.}

\item{\code{dataType}}{kind of expresses value matrix. `Counts` (Integer) or `Array` (Decimal).}

\item{\code{idType}}{kind of gene id. `ENSEMBL` or `SYMBOL`, If `ENSEMBL`, it will be automatically converted to `SYMBOL`.}

\item{\code{expMatrix}}{expresses value matrix. Should be a data.frame row named by gene ID and column named by Sample}

\item{\code{groupInfo}}{a Character Vectors ordered by samples in matrix.}

\item{\code{caseGroup}}{a Character names of case group.}

\item{\code{filterMethod}}{a function used to filter expresses value matrix. Or disable filter by set as `NULL`.}

\item{\code{matrixFiltered}}{expresses value matrix after apply \code{filterMethod}. Should be a data.frame row named by gene ID and column named by Sample}
}}

\seealso{
Other DEGContainer: 
\code{\link{MSigDB}},
\code{\link{degResults-class}}

Other DEGContainer: 
\code{\link{MSigDB}},
\code{\link{degResults-class}}
}
\concept{DEGContainer}
