% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGContainer-methods.R, R/allClass.R,
%   R/allGeneric.R, R/allMethods.R
\docType{methods}
\name{DEGContainer}
\alias{DEGContainer}
\alias{DEGContainer-class}
\alias{dataInfo}
\alias{degResults}
\alias{hyperResults}
\alias{gseResults}
\alias{MSigDB}
\alias{species<-,DEGContainer-method}
\alias{dataType<-,DEGContainer-method}
\alias{idType<-,DEGContainer-method}
\alias{expMatrix<-,DEGContainer-method}
\alias{groupInfo<-,DEGContainer-method}
\alias{caseGroup<-,DEGContainer-method}
\alias{filterMethod<-,DEGContainer-method}
\alias{matrixFiltered<-,DEGContainer-method}
\alias{geneNames<-,DEGContainer-method}
\alias{sampleNames<-,DEGContainer-method}
\alias{msigdbParam,DEGContainer-method}
\alias{msigdbGSEAparam,DEGContainer-method}
\alias{msigdbParam<-,DEGContainer-method}
\alias{msigdbGSEAparam<-,DEGContainer-method}
\alias{msigdbData,DEGContainer-method}
\alias{msigdbData<-,DEGContainer-method}
\alias{msigdbGSEAresult,DEGContainer-method}
\alias{msigdbGSEAresult<-,DEGContainer-method}
\alias{msigdbGSVAresult,DEGContainer-method}
\alias{msigdbGSVAresult<-,DEGContainer-method}
\title{Initialize an object of class \code{DEGContainer}}
\usage{
DEGContainer(
  dataInfo = NULL,
  degResults = NULL,
  hyperResults = NULL,
  gseResults = NULL,
  MSigDB = NULL
)

dataInfo(obj)

degResults(obj)

hyperResults(obj)

gseResults(obj)

MSigDB(obj)

\S4method{species}{DEGContainer}(obj) <- value

\S4method{dataType}{DEGContainer}(obj) <- value

\S4method{idType}{DEGContainer}(obj) <- value

\S4method{expMatrix}{DEGContainer}(obj) <- value

\S4method{groupInfo}{DEGContainer}(obj) <- value

\S4method{caseGroup}{DEGContainer}(obj) <- value

\S4method{filterMethod}{DEGContainer}(obj) <- value

\S4method{matrixFiltered}{DEGContainer}(obj) <- value

\S4method{geneNames}{DEGContainer}(obj, filtered = FALSE) <- value

\S4method{sampleNames}{DEGContainer}(obj, filtered = FALSE) <- value

\S4method{msigdbParam}{DEGContainer}(obj)

\S4method{msigdbGSEAparam}{DEGContainer}(obj)

\S4method{msigdbParam}{DEGContainer}(obj) <- value

\S4method{msigdbGSEAparam}{DEGContainer}(obj) <- value

\S4method{msigdbData}{DEGContainer}(obj)

\S4method{msigdbData}{DEGContainer}(obj) <- value

\S4method{msigdbGSEAresult}{DEGContainer}(obj)

\S4method{msigdbGSEAresult}{DEGContainer}(obj) <- value

\S4method{msigdbGSVAresult}{DEGContainer}(obj)

\S4method{msigdbGSVAresult}{DEGContainer}(obj) <- value
}
\arguments{
\item{dataInfo}{a dataInfo object}

\item{degResults}{a degResults object}

\item{hyperResults}{a hyperResults object}

\item{gseResults}{a gseResults object}

\item{MSigDB}{a MSigDB object}
}
\value{
a DEGContainer

The accessor functions \code{dataInfo}, \code{degResults},
  \code{hyperResults}, \code{gseResults}, \code{MSigDB} return
  the corresponding elements of a
  \code{DEGContainer}.
}
\description{
Constructs a \code{DEGContainer} object. Additional helper
  methods for manipulating \code{DEGContainer} objects are  also
  described below. We now recommend using
  \code{Create_DEGContainer} to create objects, instead.

This was the universal class for storing data and
  results. We now generally reommend using the
  \code{\link{Create_DEGContainer}} to create it.
}
\section{Slots}{

\describe{
\item{\code{dataInfo}}{store input data of workflow}

\item{\code{degResults}}{store parameters and results in \code{runDEG} module}

\item{\code{hyperResults}}{store parameters and results in \code{runHyper} module}

\item{\code{gseResults}}{store parameters and results in \code{runGSEA} module}

\item{\code{MSigDB}}{store parameters and results in \code{runMSigDB} module}
}}

