% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R, R/allMethods.R
\docType{methods}
\name{DEGContainer}
\alias{DEGContainer}
\alias{DEGContainer-class}
\alias{species<-,DEGContainer-method}
\alias{dataType<-,DEGContainer-method}
\alias{idType<-,DEGContainer-method}
\alias{expMatrix<-,DEGContainer-method}
\alias{groupInfo<-,DEGContainer-method}
\alias{caseGroup<-,DEGContainer-method}
\alias{filterMethod<-,DEGContainer-method}
\alias{matrixFiltered<-,DEGContainer-method}
\alias{geneNames<-,DEGContainer-method}
\alias{sampleNames<-,DEGContainer-method}
\alias{treatInfo,DEGContainer-method}
\alias{label,DEGContainer-method}
\alias{label_ns,DEGContainer-method}
\alias{cutFC,DEGContainer-method}
\alias{cutFDR,DEGContainer-method}
\alias{sigCol,DEGContainer-method}
\alias{sigAlpha,DEGContainer-method}
\alias{sigSize,DEGContainer-method}
\alias{sigShape,DEGContainer-method}
\alias{treatInfo<-,DEGContainer-method}
\alias{cutFC<-,DEGContainer-method}
\alias{cutFDR<-,DEGContainer-method}
\alias{label<-,DEGContainer-method}
\alias{label_ns<-,DEGContainer-method}
\alias{sigCol<-,DEGContainer-method}
\alias{sigAlpha<-,DEGContainer-method}
\alias{sigSize<-,DEGContainer-method}
\alias{sigShape<-,DEGContainer-method}
\alias{vsData,DEGContainer-method}
\alias{limma_res,DEGContainer-method}
\alias{edgeR_res,DEGContainer-method}
\alias{DESeq2_res,DEGContainer-method}
\alias{merge_res,DEGContainer-method}
\alias{vsData<-,DEGContainer-method}
\alias{DESeq2_res<-,DEGContainer-method}
\alias{limma_res<-,DEGContainer-method}
\alias{edgeR_res<-,DEGContainer-method}
\alias{merge_res<-,DEGContainer-method}
\alias{hyperKEGGparam,DEGContainer-method}
\alias{hyperKEGGparam<-,DEGContainer-method}
\alias{hyperGOparam,DEGContainer-method}
\alias{hyperGOparam<-,DEGContainer-method}
\alias{hyperRes,DEGContainer-method}
\alias{hyperRes<-,DEGContainer-method}
\alias{gseKEGGparam,DEGContainer-method}
\alias{gseKEGGparam<-,DEGContainer-method}
\alias{gseGOparam,DEGContainer-method}
\alias{gseGOparam<-,DEGContainer-method}
\alias{gseRes,DEGContainer-method}
\alias{gseRes<-,DEGContainer-method}
\alias{msigdbParam,DEGContainer-method}
\alias{msigdbGSEAparam,DEGContainer-method}
\alias{msigdbHyperParam,DEGContainer-method}
\alias{msigdbParam<-,DEGContainer-method}
\alias{msigdbGSEAparam<-,DEGContainer-method}
\alias{msigdbHyperParam<-,DEGContainer-method}
\alias{msigdbData,DEGContainer-method}
\alias{msigdbData<-,DEGContainer-method}
\alias{msigdbGSEAresult,DEGContainer-method}
\alias{msigdbGSEAresult<-,DEGContainer-method}
\alias{msigdbHyperResult,DEGContainer-method}
\alias{msigdbHyperResult<-,DEGContainer-method}
\alias{msigdbGSVAresult,DEGContainer-method}
\alias{msigdbGSVAresult<-,DEGContainer-method}
\alias{msigdbTreat,DEGContainer-method}
\alias{msigdbTreat<-,DEGContainer-method}
\title{Class \code{DEGContainer}}
\usage{
\S4method{species}{DEGContainer}(obj) <- value

\S4method{dataType}{DEGContainer}(obj) <- value

\S4method{idType}{DEGContainer}(obj) <- value

\S4method{expMatrix}{DEGContainer}(obj) <- value

\S4method{groupInfo}{DEGContainer}(obj) <- value

\S4method{caseGroup}{DEGContainer}(obj) <- value

\S4method{filterMethod}{DEGContainer}(obj) <- value

\S4method{matrixFiltered}{DEGContainer}(obj) <- value

\S4method{geneNames}{DEGContainer}(obj, filtered = FALSE) <- value

\S4method{sampleNames}{DEGContainer}(obj, filtered = FALSE) <- value

\S4method{treatInfo}{DEGContainer}(obj)

\S4method{label}{DEGContainer}(obj)

\S4method{label_ns}{DEGContainer}(obj)

\S4method{cutFC}{DEGContainer}(obj)

\S4method{cutFDR}{DEGContainer}(obj)

\S4method{sigCol}{DEGContainer}(obj)

\S4method{sigAlpha}{DEGContainer}(obj)

\S4method{sigSize}{DEGContainer}(obj)

\S4method{sigShape}{DEGContainer}(obj)

\S4method{treatInfo}{DEGContainer}(obj) <- value

\S4method{cutFC}{DEGContainer}(obj) <- value

\S4method{cutFDR}{DEGContainer}(obj) <- value

\S4method{label}{DEGContainer}(obj) <- value

\S4method{label_ns}{DEGContainer}(obj) <- value

\S4method{sigCol}{DEGContainer}(obj) <- value

\S4method{sigAlpha}{DEGContainer}(obj) <- value

\S4method{sigSize}{DEGContainer}(obj) <- value

\S4method{sigShape}{DEGContainer}(obj) <- value

\S4method{vsData}{DEGContainer}(obj)

\S4method{limma_res}{DEGContainer}(obj)

\S4method{edgeR_res}{DEGContainer}(obj)

\S4method{DESeq2_res}{DEGContainer}(obj)

\S4method{merge_res}{DEGContainer}(obj)

\S4method{vsData}{DEGContainer}(obj) <- value

\S4method{DESeq2_res}{DEGContainer}(obj) <- value

\S4method{limma_res}{DEGContainer}(obj) <- value

\S4method{edgeR_res}{DEGContainer}(obj) <- value

\S4method{merge_res}{DEGContainer}(obj) <- value

\S4method{hyperKEGGparam}{DEGContainer}(obj)

\S4method{hyperKEGGparam}{DEGContainer}(obj) <- value

\S4method{hyperGOparam}{DEGContainer}(obj)

\S4method{hyperGOparam}{DEGContainer}(obj) <- value

\S4method{hyperRes}{DEGContainer}(obj)

\S4method{hyperRes}{DEGContainer}(obj) <- value

\S4method{gseKEGGparam}{DEGContainer}(obj)

\S4method{gseKEGGparam}{DEGContainer}(obj) <- value

\S4method{gseGOparam}{DEGContainer}(obj)

\S4method{gseGOparam}{DEGContainer}(obj) <- value

\S4method{gseRes}{DEGContainer}(obj)

\S4method{gseRes}{DEGContainer}(obj) <- value

\S4method{msigdbParam}{DEGContainer}(obj)

\S4method{msigdbGSEAparam}{DEGContainer}(obj)

\S4method{msigdbHyperParam}{DEGContainer}(obj)

\S4method{msigdbParam}{DEGContainer}(obj) <- value

\S4method{msigdbGSEAparam}{DEGContainer}(obj) <- value

\S4method{msigdbHyperParam}{DEGContainer}(obj) <- value

\S4method{msigdbData}{DEGContainer}(obj)

\S4method{msigdbData}{DEGContainer}(obj) <- value

\S4method{msigdbGSEAresult}{DEGContainer}(obj)

\S4method{msigdbGSEAresult}{DEGContainer}(obj) <- value

\S4method{msigdbHyperResult}{DEGContainer}(obj)

\S4method{msigdbHyperResult}{DEGContainer}(obj) <- value

\S4method{msigdbGSVAresult}{DEGContainer}(obj)

\S4method{msigdbGSVAresult}{DEGContainer}(obj) <- value

\S4method{msigdbTreat}{DEGContainer}(obj)

\S4method{msigdbTreat}{DEGContainer}(obj) <- value
}
\value{
The accessor functions \code{dataInfo}, \code{degResults},
  \code{hyperResults}, \code{gseResults}, \code{MSigDB} return
  the corresponding elements of a
  \code{DEGContainer}.
}
\description{
This was the universal class for storing data and
  results. We now generally reommend using the
  \code{\link{Create_DEGContainer}} to create it.
}
\section{Slots}{

\describe{
\item{\code{dataInfo}}{store input data of workflow}

\item{\code{degResults}}{store parameters and results in \code{runDEG} module}

\item{\code{hyperResults}}{store parameters and results in \code{runHyper} module}

\item{\code{gseResults}}{store parameters and results in \code{runGSEA} module}

\item{\code{MSigDB}}{store parameters and results in \code{runMSigDB} module}
}}

