% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_DEG.R
\name{DEGtopHeatmap}
\alias{DEGtopHeatmap}
\title{Heatmap for DEG data frame}
\usage{
DEGtopHeatmap(
  object,
  which,
  top = 50,
  filename = NA,
  show_gene = TRUE,
  category = "H",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2],
  ...
)
}
\arguments{
\item{object}{a counts data frame of rows in genes and columns in samples}

\item{which}{which model of deg analysis. kinds of DEG; can be "limma", "edgeR", "DESeq2" or "MSigDB"}

\item{top}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{filename}{NA or a file path}

\item{show_gene}{logical,show gene name}

\item{category}{MSigDB collection abbreviation, such as H or C1.}

\item{palette}{a color palette for plots}

\item{...}{More \code{\link[pheatmap]{pheatmap}} parameters.}
}
\value{
a heatmap plot file
}
\description{
default will return a top 100 deg heatmap in p value = 0.05
}
\examples{
DEGtopHeatmap(object,which = "limma")
}
