% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-class.R, R/results-methods.R
\docType{methods}
\name{MSigDB}
\alias{MSigDB}
\alias{MSigDB-class}
\alias{msigdbParam,MSigDB-method}
\alias{msigdbGSEAparam,MSigDB-method}
\alias{msigdbParam<-,MSigDB-method}
\alias{msigdbGSEAparam<-,MSigDB-method}
\alias{msigdbData,MSigDB-method}
\alias{msigdbData<-,MSigDB-method}
\alias{msigdbGSEAresult,MSigDB-method}
\alias{msigdbGSEAresult<-,MSigDB-method}
\alias{msigdbGSVAresult,MSigDB-method}
\alias{msigdbGSVAresult<-,MSigDB-method}
\title{Class \code{MSigDB}}
\usage{
\S4method{msigdbParam}{MSigDB}(obj)

\S4method{msigdbGSEAparam}{MSigDB}(obj)

\S4method{msigdbParam}{MSigDB}(obj) <- value

\S4method{msigdbGSEAparam}{MSigDB}(obj) <- value

\S4method{msigdbData}{MSigDB}(obj)

\S4method{msigdbData}{MSigDB}(obj) <- value

\S4method{msigdbGSEAresult}{MSigDB}(obj)

\S4method{msigdbGSEAresult}{MSigDB}(obj) <- value

\S4method{msigdbGSVAresult}{MSigDB}(obj)

\S4method{msigdbGSVAresult}{MSigDB}(obj) <- value
}
\value{
The accessor functions \code{msigdbParam}, \code{msigdbData},
  \code{msigdbGSEAparam}, \code{msigdbGSEAresult}, \code{msigdbGSVAresult}
  return the corresponding elements of a
  \code{DEGContainer} or \code{MSigDB}.
}
\description{
This was the class for storing data from MSigDB and analysis
  results.
  We now generally recommend using the \code{\link{Create_DEGContainer}}
  to create it in \code{DEGContainer} obj.
}
\section{Slots}{

\describe{
\item{\code{msigdbParam}}{list_OR_NULL.}

\item{\code{msigdbData}}{list.}

\item{\code{msigdbGSEAparam}}{list_OR_NULL.}

\item{\code{msigdbGSEAresult}}{list.}

\item{\code{msigdbGSVAresult}}{list.}
}}

